[CARBONDATA-3910] Fix load failure in cluster when csv present in local file system in case of global sort
Why is this PR needed?
when the csv file is present in local file system and we load the table in cluster, it fails with file not found.
This is because when we try to update the file path, we remove one slash in schema which causes this issue.

What changes were proposed in this PR?
Its better to check for "file:/" instead of "file:///" for local file, as getUpdatedfilePath API may remove
one slash and cause this issue.

This closes #3838