[CARBONDATA-3871][FOLLOW-UP] Fix memory issue of loading sort step when get data from row page
Why is this PR needed?
PR #3804 free the row page after running in-memory intermeidate merge, but the data will still be used in final sort because that merge only acquire the memory address and row page index instead of moving the rows from here to there.

What changes were proposed in this PR?
Remove code for freeing the row page in case of in-memory intermeidate merge.
As a special case, add comment for explanation.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3845