[CARBONDATA-3918] Fix extra data in count(*) after multiple updates and index server running
Why is this PR needed?
Select count * gives extra data after multiple updates with the index server running.
This is because, once the horizontal compaction happens, it stores the index fils to cache
and create new index and data files, so if the table is updated or deleted table, we will
exclude those splits after getting all splits. Since once the splits come from index server
since loadmetadatadetails are transient in Segment object, we will have null value for it as
the slits are serialized from index server. Because of which it won't be able to filter out
the IUD old segments. So it leads to extra data in count *.

What changes were proposed in this PR?
Once we get the splits from the index server, then from the validSegments, get the
loadmetadataDetails and readCommittedScope and set into the splits which solve this problem.

This closes #3853