[CARBONDATA-3928] Handled the Strings which length is greater than 32000 as a bad record.
Why is this PR needed?
Currently, when the string length exceeds 32000 then the load is failed. The particular
record should be handled as a bad record and load should not be failed.

What changes were proposed in this PR?
At place of throwing exception in case of strings greater than 32000 bytes now we will
handle as a bad record and load will not fail.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3865