[CARBONDATA-3956] Reindex command on SI table
Why is this PR needed?
In the main table with SI tables after every
load/insert command , SILoadEventListenerForFailedSegments.scala
checks for missing segments or segments mismatch in SI table
and loads the missing/deleted segments to the SI table.
This functionality is dependent on the load/insert command.
A new functionality/command is required so that this repair
logic can be made independent to the load/insert logic.

What changes were proposed in this PR?
Added a separate SQL reindex command(reindex [index_table]
on table maintable) to call the SI repair logic without load/insert.

This closes #3873