[CARBONDATA-3940] Remove "spark.sql.sources.commitProtocolClass" configuration from SparkCarbonTableFormat
Why is this PR needed?
During the load process, commitTask fails with high probability.
The exception stack shows that it was throwed by HadoopMapReduceCommitProtocol, not CarbonSQLHadoopMapMapReduceCommitProtocol,
implying that there is class init error during the initializing of "Committer".
It should have been initialized as CarbonSQLHadoopMapMapReduceCommitProtocol, but was incorrectly initialized to HadoopMapReduceCommitProtocol.

What changes were proposed in this PR?
Init the committer to be CarbonSQLHadoopMapMapReduceCommitProtocol directly

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3883