[CARBONDATA-3963]Fix hive timestamp data mismatch issue and empty data during query issue
Why is this PR needed?
1. when carbon table is queried through beeline, gives empty data.
2. the timestamp data given by carbon table is different than hive table.
Because during read we used to send the Timestamp value to WritableTimestampObjectInspector's
create method, which basically was converting the value to EpochMillli, which resulted in the
time value go different than the actual input data.

What changes were proposed in this PR?
1. Pass the carbon storage_handler into the table parameters
2. Convert the the long value to formatted string and directly form the TimestampWritableV2,
by passing the Hive's Timestamp object which avoid EpochMilli and gives the proper data as output.

This closes #3903