[CARBONDATA-3967] cache partition on select to enable faster pruning
Why is this PR needed?
spark's getPartition method can sometime take a lot of time which can degrade the performance for the queries.

What changes were proposed in this PR?
Cache the partition on each load and query so that the subsequent queries can perform faster. Caching is done under the expiring map interface so the cache expiration time can work for this also.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3908