[CARBONDATA-3975]Fix wrong data from carbondata for binary column when read via hive
Why is this PR needed?
When binary data read via hive, the results are not the actual
inserted data. This is because the BytesWritable objected will be
given for binary data, we were doing the toString() to get the data
object which gives wrong data for binary as it gives objects to string.

What changes were proposed in this PR?
For BytesWritable, get the bytes from the object and get the bytes
only for the specific length in object and not the extra bits.

This closes #3915