[CARBONDATA-3974] Improve partition purning performance in presto carbon integration
Why is this PR needed?
a) For 200K segments table in cloud, presto partition query was taking
more than 5 hours. the reason is it was reading all segment files for
partition pruning. Now it is less than a minute !
b) If the query filter is < or > or more than one value.
Partition pruning is not working in presto.
c) prestodb profile compilation is broken from previous PR

What changes were proposed in this PR?
a) HiveTableHandle already have partition spec, matching for the filters
(it has queried metastore to get all partitions and pruned it). So, create
partitionSpec based on that. Also handled for both prestodb and prestosql
b) #3885 , broke prestodb compilation, only prestosql is compiled.
c) #3887, also didn't handled prestodb

This closes #3913