[CARBONDATA-3984] compaction on table having range column after altering datatype from string to long string fails
Why is this PR needed?
Fix multiple issues occurred after altering the column dataType to long_string_column.
a) When dataType of a String column which is also provided as range column in table properties
is altered to long_string_column. It throws following error while performing compaction on the
table.
VARCHAR not supported for the filter expression;
b) When longStringColumn name is present in UpperCase then Validation check for SortColumns
fails. As names of SortColumns are present in LowerCase.

What changes were proposed in this PR?
a) Added condition for VARCHAR datatype in all conditional expressions.
b) Changed the long_string_columns name to lowerCase before doing validation.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3923