[CARBONDATA-3988] Allow SI creation on first dimension column
Why is this PR needed?
In some scenarios like, A table having 100 columns. with sort columns as (A,B,C). Currently we don't allow SI creation on column 'A'. Due to this If the filter is B=x, A=y. There will be huge task to scan main table as SI is not supported for column 'A'. Instead if we allow SI for A,B. There will be less task to scan main table and over all query execution will be faster.
Let user decide he need to create SI for first column or not, instead of we blocking always.

What changes were proposed in this PR?
Allow first dimension column also as SI column.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3924