[CARBONDATA-4036]Fix special char(`) issue in create table, when column name contains ` character
Why is this PR needed?
When a table is created with special character ``` , it fails when
parsing to create field object. This is because ` character is used for
tokenizing while parsing, and it will use the same char present in column
name for parsing and it leads to wrong field object creation.

What changes were proposed in this PR?
when the column name contains the ``` , just for parsing , replace the
character with ! character, then once the parsing is done and the files
object is created, then replace again with the original one. This solution
is considered so as to avoid the changes in parsing layer, as it might
induce other layers of parsing. So this solution is considered.

This closes #3983