[CARBONDATA-3914] Fixed exception on reading data from carbon-hive empty table.
Why is this PR needed?
Reading data from empty carbontable through hive beeline was giving 'Unable read Carbon Schema' exception when no data is present in the carbon table. In case of empty table, some additional fields are present in the columns. And also schema is comma-separated which we are expecting to be colon-separated.

What changes were proposed in this PR?
Return the empty data if no data is present. Validated the columns and column types and removed, if any extra field was added. Also parsed the schema in such a way that it can consider comma-separated values also.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3936