[CARBONDATA-3995] Support presto querying older complex type stores
Why is this PR needed?
Before carbon 2.0, complex child length is stored as SHORT for string,
varchar, binary, date, decimal types.
So, In 2.0 as it is stored as INT, presto complex query code always
assumes it as INT and goes to out of bound exception when old store is queried.

What changes were proposed in this PR?
If INT_LENGTH_COMPLEX_CHILD_BYTE_ARRAY encoding is present, parse as
INT, else parse as SHORT. so, that both stores can be queried.

This closes #3937