[CARBONDATA-3998] Fixed file not found exception for hive during insert
Why is this PR needed?
FileNotFoundException being thrown in hive during insert.

What changes were proposed in this PR?
During insert the path in the parameters for write is for staging directory,
because carbon is non-native table therefore this directory does not exist.
When getAbsolutePath is called on this directory, it throws FileNotFound
exception. Removed the getAbsolutePath call to avoid this.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3941