[CARBONDATA-4001] Fix SI global sort load on partition table
Why is this PR needed?
On a partition table, when SI is created with global sort and
when data is loaded. It shows 0 rows as main table query results is 0 rows.

What changes were proposed in this PR?
For partition table, local sort SI flow current.segmentfile
is set in CarbonSecondaryIndexRDD
For the global sort, this value was not set. so, the main table
query was resulting in 0 rows. Setting this value for global sort flow also.

This closes #3944