[CARBONDATA-4015] Remove hard-coded parameters of the ICarbonLock.lockWithRetries method
Why is this PR needed?
Retrycount and retryInterval of compactLock and updateLock is fixed as "3" when they try to get lock, default values of these two parameters is exactly "3", therefore, it is not neccessary to be fixed in code, it is also good to allow users to configure them in the carbon.properties by removing "3".

What changes were proposed in this PR?
Removed the parameters of compactLock and updateLock when they try to get lock.
Besides, I made some obvious but trivial changes.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3964