[CARBONDATA-4027] Fix the wrong modifiedtime of loading files in insert stage
Why is this PR needed?
In the insertstage flow, there is a empty file with suffix '.loading' to mark the stage in the status of 'in processing'. We update the modifiedtime of '.loading' file for monitoring the insertstage start time, which can be used for calculate TIMEOUT, help to retry and recovery.
Before, we use setModifiedTime function to update the modifiedtime, which has a serious bug.
For S3 file, setModifiedTime operation do not take effect. leading to the incorrect inserstage starttime of 'loading' file.

What changes were proposed in this PR?
Update the modifiedtime of loading files based on recreating files.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No