[CARBONDATA-4027] Fix the wrong modifiedtime of loading files in insert stage
Why is this PR needed?

ISSUE1: In the insertstage flow, there is a empty file with suffix '.loading'
to mark the stage in the status of 'in processing'. We update the modifiedtime
of '.loading' file for monitoring the insertstage start time, which can be used
for calculate TIMEOUT, help to retry and recovery. Before, we use setModifiedTime
function to update the modifiedtime, which has a serious bug. For S3 file,
setModifiedTime operation do not take effect. leading to the incorrect
inserstage starttime of 'loading' file.

ISSUE2: For now, Insertstage non-parttion table will not merge index files,
which will degrade the query performance heavily.

What changes were proposed in this PR?
1. Update the modifiedtime of loading files based on recreating files.
2. Trigger indexfiles merging after insertstage non-parttiontable.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3977