[CARBONDATA-4034] Improve the time-consuming of Horizontal Compaction for update
Why is this PR needed?
The horizontal compaction flow will be too slow when updating with lots of segments(or lots of blocks), so we try to analyze and optimize it for time-consuming problem.

What changes were proposed in this PR?
1. In performDeleteDeltaCompaction, optimize the method getSegListIUDCompactionQualified.
2. Combine two traversals of segments which have same process.
3. Use forming delete delta file name instead of listFiles operation.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3986