[CARBONDATA-4043] Fix data load failure issue for columns added in legacy store
Why is this PR needed?
When dimension is added in older versions like 1.1, by default it will be sort column.
In sort step we assume data will be coming as sort column in the beginning.
But the added column will be at last eventhough sort column. So, while building
the dataload configurations for loading data, we rearrange the columns(dimensions
and datafields) in order to bring the sort column to beginning and no-sort to last
and revert them back to schema order before FinalMerge/DataWriter step.

Issue:
Data loading is failing because of castException in data writing step in case
of NO_SORT and in final sort step in case of LOCAL_SORT.
NO_SORT:
In this flow, datafields are rearranged based on sort columns order, which is not required.
LOCAL_SORT:
During FinalMerge, we assume intermediate row data will have only sort column
data in schema order. But intermediate sort row data can contain no-dictionary
and dictionary data also. So, the information about the actual index of sort column data is insufficient.

What changes were proposed in this PR?
NO_SORT:
Update datafields based on sort columns only if the table is a sort table and sort columns are configured.
LOCAL_SORT:
Collect the sort column info from all the dimensions into a map and identify the sort column index using it and compare only sort column data in final merge.

This closes #3995