[CARBONDATA-4022] Fix invalid path issue for segment added through alter add segment query
Why is this PR needed?
Segment Added through alter table add segment query doesn't require tablepath in its
filepath, because the segment is present at some external location which is not inside
the table.

What changes were proposed in this PR?
Filepath for normal segments (which are not added through alter table add segment query)
is tablePath + filePath. But for segments added through alter table add segment query
doesn't require tablePath, so we directly use the filePath of the segment.

This closes #4017