[CARBONDATA-4054] Support data size control for minor compaction
Why is this PR needed?
Currentlly, minor compaction only consider the num of segments and major
compaction only consider the SUM size of segments, but consider a scenario
that the user want to use minor compaction by the num of segments but he
dont want to merge the segment whose datasize larger the threshold for
example 2GB, as it is no need to merge so much big segment and it is time
costly.

What changes were proposed in this PR?
add a parameter to control the threshold of segment included
in minor compaction, so that the user can specify the segment not included
in minor compaction once the datasize exeed the threshold, system level and table level can be set, and if not set the use default
value.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #4020