[HOTFIX] Refact Carbon Util
Why is this PR needed?
Currentlly, we have some Carbon{$FUNCTION_NAME}Util, for example CarbonSparkUtil, CarbonQueryUtil, CarbonMergeUtil, CarbonLoadUtil, and we also have some CarbonUtil/CarbonUtils which has some mixed functions in, we should clean code and move the functions in CarbonUtils to the specified Util where it should be.

What changes were proposed in this PR?
Refact the code to clean it

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #4029