[CARBONDATA-4076] Fix MV having Subquery alias used in query projection #4038
Why is this PR needed?
If alias is used in subquery and that alias is used in main query projection,
then the query is not hitting MV after creation.

For example:

Select a, sum(b) from (select c, d b from t1)

This is because, currently, we are not checking for subquery alias child node
for matching the query.

What changes were proposed in this PR?
1. If the outputList contains Alias, then compare the alias child of subsume/subsumer
2. In GroupbyNoselect, compare the alias.sql subsume/subsumer
3. Added logs for MV Rewrite and matching

This closes #4038