[CARBONDATA-4078] Add external segment and query with index server fails
Why is this PR needed?
Query after adding an external segment to carbon table tries to getSplits from Index server
and throws an exception as it cannot read the external(orc/parquet) file format.
When the fallback mode is disabled, it throws an exception and fails.

What changes were proposed in this PR?
To avoid the exception, filtered only valid carbon segments to cache in the index server.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No, tested in cluster.

This closes #4047