[CARBONDATA-4094]: Fix fallback count(*) issue on partition table with index server
Why is this PR needed?
The used asJava converts to java "in place", without copying the whole data to save
time and memory and it just simply wraps the scala collection with a class that
conforms to the java interface and thus java serializer is not able to serialize it.

What changes were proposed in this PR?
Converting it to list, so that it is able to serialize a list.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #4061