[CARBONDATA-4089] Create table with location, if the location doesn't have scheme, the default will be local file system, which is not the file system defined by fs.defaultFS
Why is this PR needed?
Create table with location, if the location doesn't have scheme, the default will be local file system, which is not the file system defined by fs.defaultFS.

What changes were proposed in this PR?
If the location doesn't have scheme, add the fs.defaultFS scheme to the beginning of the location.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #4065