[CARBONDATA-4099] Fixed select query on main table with a SI table in case of concurrent load, compact and clean files operation
Why is this PR needed?
There were 2 issues in the clean files post event listener:

1. In concurrent cases, while writing entry back to the table status file, wrong path was given,
due to which table status file was not updated in the case of SI table.
2. While writing the loadmetadetails to the table status file during concurrent scenarios,
we were only writing the unwanted segments and not all the segments, which could make segments
stale in the SI table
Due to these 2 issues, when selet query is executed on SI table, the tablestatus would have entry
for a segment but it's carbondata file would be deleted, thus throwing an IO Exception.
3. Segment ID is null when writing hive table

What changes were proposed in this PR?
1.& 2. Added correct table status path as well sending the correct loadmetadatadetails to be updated in
the table status file. Now when select query is fired on the SI table, it will not throw
carbondata file not found exception
3. set the load model after setup job of committer

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #4066