[CARBONDATA-4145] Query fails and the message "File does not exist: xxxx.carbondata" is displayed
Why is this PR needed?
If an exception occurs when the refresh index command is executed, a task has been successful. The new query will be failed.
Reason: After the compaction task is executed successfully, the old carbondata files are deleted. If other exception occurs, the deleted files are missing.
This PR will fix this issue.

What changes were proposed in this PR?
When all tasks are successful, the driver deletes the old carbondata files.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No